requirements.txt
feedparser==6.0.11
requests==2.31.0
anthropic==0.25.1
python-telegram-bot==20.7
python-dotenv==1.0.1
